+ Clean up warnings issued by gcc when --enable-all-warnings is   specified on the configure line.   - Use static_cast to eliminate warning about comparing signed and     unsigned integer types.   - Change some variables from 'int' to 'size_t.'   - Initialize all member data from the constructor initializer list     (using default values).   - Update indentation to match Draco standard style and untabify     source code.

+++ src/RTT_Format_Reader/CellDataIDs.cc
@@ -49,11 +49,11 @@ void CellDataIDs::readData(ifstream & meshfile)
     int dataIDNum;
     string dummyString;
 
-    for (unsigned i = 0; i < dims.get_ncell_data(); ++i)
+    for (size_t i = 0; i < static_cast<size_t>(dims.get_ncell_data()); ++i)
     {
 	Check(i<names.size() && i<units.size());
 	meshfile >> dataIDNum >> names[i] >> units[i];
-	Insist(dataIDNum == i+1,
+	Insist(static_cast<size_t>(dataIDNum) == i+1,
 	       "Invalid mesh file: cell data ID out of order");
 	std::getline(meshfile, dummyString);
     }
